Don't change directory anymore in scansys.py.
Instead, track compiler output and clean up only the files that should
have been created.  This forces us to stop pretending cl's command line
arguments work the same way as everyone else's.  They don't.

All compilers-which-are-not-MSVC are now being invoked with -S instead
of -c, eliminating the need for the special :compile_opt tag applied to
clang-for-android.  Hopefully this works with all the wacky proprietary
Unix compilers as well.