Restore a previously-discovered presence test subtlety.
On some systems (notably IRIX6 and OSX) there exist headers that
require their dependencies to _preprocess_ without error, and/or
only preprocess without error in conformance mode.  Thus, if a
failure of the presence test is not due to nonexistence, we should
mark the header as PRESENT, not BUGGY.  If it is actually BUGGY, we
will discover this during the dependencies test.

Also add another header that was only mentioned in prereqs.ini to
headers.ini, and make ucontext_deprecated a caution, not a full error.