Nuke the conform/thread booleans being passed around everywhere.
There is now a CompilationMode object, which is currently just
a wrapper around those same two booleans, but will become more
capable in the future.  More importantly, the mode is now a
property of the *dataset*, rather than each Header object carrying
info for all N modes.  This will make output hairier (and may well
be the final nail in the coffin of the current output format) but
it's worth it for how much tidier the code performing the actual
tests is.