Run tests in all relevant modes, again.
With the new setup, this means we are probing dependencies in all
modes, and that in turn means we have to allow for headers that are
buggy (most commonly suffering from 'legacy_type_decls') as hard
dependencies of other headers. Supersedes an ad-hoc patch on one
such situation in errors.ini.