Implement probe_max_posix_level, finally.
Uses a heuristic which so far has only been tested on OSX, but was designed
based on knowledge of several other systems.  It too has the possibility of
throwing a NoSuchHeaderError to tell the caller to forget about this whole
class of tests, so requiring some changes to the caller as well.