Only use `long long` and `long double` in content tests when truly necessary.
The default mode for many compilers includes warnings for any use of
`long long`, which we turn into errors.  It's fine for this to cause
C99 functions with `long long` in their official signature to be
unusable in C89 modes, but it shouldn't affect things like `snprintf`
which can be used without `long long`.

Also give this treatment to `long double`, even though it *was* in C89
and I haven't yet tripped over a compiler that complains about it,
because it wasn't thoroughly integrated into the library before C99.